refactor: create logOutModal (#101)
